removed unused imports, cleaned up code
